add: 요청 친구 코드 API 추가
